move autoJump to separate file
